requirements.txt
asgiref
certifi==2022.6.15
charset-normalizer==2.0.12
Django
gunicorn
idna==3.3
Pillow==8.0.1
requests
sqlparse==0.4.2
urllib3==1.26.9
ibmcloudant
